fix(plrt.py): reg/tree vars conversion to indices
Regression and tree variables are converted to indices of self.feats when the model is first initialized. After that, they are passed as to child nodes as those indices. not as the original values passed to the init method. The _var_iindices_already boolean flag  is added to account for this difference.